requirements.txt
# Graphics
pygame
moderngl

# Calculations
numpy
pandas
PyGLM